Charging Rhino
The RhinoEnemy class was created, and it implements its charging attack toward the player.

If the player is in range, it charges towards the player, following the values from an AnimationCurve for tweakibility.

Once it has charged, it cools down for a brief period before charging again.

The EnemyBase class was expanded with further features that will be present in the other enemies - I'm sure there will be much more to come as we develop them further!