Consistency amongst the Enemies and QOL changes
The same naming is now used amongst all the enemies

OnDeath() is going to do the same thing for every enemy, so removed it from being pure virtual, into just a base class member function

Revealed the FacingDirection variable to the Editor, so that can be set by the designers

In the Rhino and Wasp, I moved each of the states into their own functions for better readibility. Also wrapped the changing of their states in a Set() function rather than a property to get rid of redundancies, and allow for changes to be tracked in the callstack.

 Some Debug messages were removed from the Ant class