Camera zooming!
The CameraController class was created to live on a cinemachine virtual camera. This will be expanded to set the various parameters we need, but at the moment only adjusts zoom. Lerp was used to make it a smooth transition over time.

The CameraTrigger objects will exist around the scene to trigger these camera effects. It has a bunch of public fields, so the designers can edit to their heart's content. When colliding with the player, it tells the CameraController what to do